+++ ucp/_libs/__init__.py
@@ -1,0 +1,4 @@
+# Copyright (c) 2019-2021, NVIDIA CORPORATION. All rights reserved.
+# See file LICENSE for terms.
+
+from .utils import nvtx_annotate  # noqa
