+++ ucp/_libs/ucx_rkey.pyx
@@ -1,4 +1,5 @@
-# Copyright (c) 2021       UT-Battelle, LLC. All rights reserved.
+# Copyright (c) 2021, NVIDIA CORPORATION. All rights reserved.
+# Copyright (c) 2021, UT-Battelle, LLC. All rights reserved.
 # See file LICENSE for terms.
 
 # cython: language_level=3
