+++ ucp/_libs/ucxio.pyx
@@ -1,3 +1,6 @@
+# Copyright (c) 2021, UT-Battelle, LLC. All rights reserved.
+# See file LICENSE for terms.
+
 from io import RawIOBase
 
 from .arr cimport Array
