+++ ucp/_libs/utils.pyx
@@ -4,6 +4,12 @@
 
 # cython: language_level=3
 
+import fcntl
+import glob
+import os
+import socket
+import struct
+
 from cpython.buffer cimport PyBUF_FORMAT, PyBUF_ND, PyBUF_WRITABLE
 from libc.stdio cimport (
     FILE,
@@ -18,9 +24,8 @@ from libc.stdio cimport (
 )
 from libc.stdlib cimport free
 
+from .exceptions import UCXConfigError, UCXError
 from .ucx_api_dep cimport *
-
-from ..exceptions import UCXConfigError, UCXError
 
 
 cdef FILE * create_text_fd():
@@ -182,3 +187,64 @@ def is_am_supported():
         return get_ucx_version() >= (1, 11, 0)
     ELSE:
         return False
+
+
+def get_address(ifname=None):
+    """
+    Get the address associated with a network interface.
+
+    Parameters
+    ----------
+    ifname : str
+        The network interface name to find the address for.
+        If None, it uses the value of environment variable `UCXPY_IFNAME`
+        and if `UCXPY_IFNAME` is not set it defaults to "ib0"
+        An OSError is raised for invalid interfaces.
+
+    Returns
+    -------
+    address : str
+        The inet addr associated with an interface.
+
+    Examples
+    --------
+    >>> get_address()
+    '10.33.225.160'
+
+    >>> get_address(ifname='lo')
+    '127.0.0.1'
+    """
+
+    def _get_address(ifname):
+        ifname = ifname.encode()
+        with socket.socket(socket.AF_INET, socket.SOCK_DGRAM) as s:
+            return socket.inet_ntoa(
+                fcntl.ioctl(
+                    s.fileno(), 0x8915, struct.pack("256s", ifname[:15])  # SIOCGIFADDR
+                )[20:24]
+            )
+
+    def _try_interfaces():
+        prefix_priority = ["ib", "eth", "en"]
+        iftypes = {p: [] for p in prefix_priority}
+        for i in glob.glob("/sys/class/net/*"):
+            name = i.split("/")[-1]
+            for p in prefix_priority:
+                if name.startswith(p):
+                    iftypes[p].append(name)
+        for p in prefix_priority:
+            iftype = iftypes[p]
+            iftype.sort()
+            for i in iftype:
+                try:
+                    return _get_address(i)
+                except OSError:
+                    pass
+
+    if ifname is None:
+        ifname = os.environ.get("UCXPY_IFNAME")
+
+    if ifname is not None:
+        return _get_address(ifname)
+    else:
+        return _try_interfaces()
