+++ ucp/_libs/utils_test.py
@@ -1,3 +1,6 @@
+# Copyright (c) 2019-2021, NVIDIA CORPORATION. All rights reserved.
+# See file LICENSE for terms.
+
 import multiprocessing as mp
 
 from ucp._libs import ucx_api
