+++ ucp/endpoint_reuse.py
@@ -1,3 +1,6 @@
+# Copyright (c) 2019-2021, NVIDIA CORPORATION. All rights reserved.
+# See file LICENSE for terms.
+
 import ctypes
 import pickle
 import uuid
